fix(factory-graph): keep authored-size validation within lint limits
